+++ templater/lang/en/lang.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * English language file for template plugin
+ *
+ * @author Daniel Dias Rodrigues <danieldiasr@gmail.com>
+ */
+$lang['template_not_found'] = 'Template "%1$s" not found. ';
+$lang['click_create'] = '[Click here to create it]';
+$lang['click_edit'] = '[Click here to edit it]';
+$lang['stopped_recursion'] = 'Processing of template "%1$s" stopped due to recursion. ';
+$lang['no_permissions_view'] = '<div class="template"> No permissions to view the template </div>';