+++ templater/lang/pt-br/lang.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * Brazilian portuguese language file for template plugin
+ *
+ * @author Daniel Dias Rodrigues <danieldiasr@gmail.com>
+ */
+$lang['template_not_found'] = 'O template "%1$s" não foi encontrado. ';
+$lang['click_create'] = '[Clique aqui para criá-lo]';
+$lang['click_edit'] = '[Clique aqui para editá-lo]';
+$lang['stopped_recursion'] = 'O processamento do template "%1$s" foi interrompido devido à recursão. ';
+$lang['no_permissions_view'] = '<div class="template"> Sem permissões para visualizar o template </div>';